fixes problems with gzip decoding
I'm unsure why this hasn't been reported as a bug for other users, but gzipped responses were not being decoded which was crashing the deserializer. upon clearing that up, there was a problem with the indexing in the constructor for DownloadedChunkRowSet, so I removed that feature. For some reason this is also running a lot faster, not sure why.